// src/utils/interfaces/authors/get-all-authors.types.ts
export interface GetAllAuthorsResponse {
	id: { type: string; value: string };
	name: { type: string; value: string };
	nip: { type: string; value: string };
	image: { type: string; value: string };
	email: { type: string; value: string };
	expertises: { type: string; value: string };
	department: { type: string; value: string };
}